heading is now a ushort. various other logic fixes for projectiles

+++ src/conversation/conversation_functions/x_obj_stuff.cs
@@ -68,7 +68,7 @@ namespace Underworld
                     {
                         if (look.CanBeIdentified(obj))
                         {
-                            obj.heading = identifyObject;
+                            obj.heading = (ushort)identifyObject;
                         }
                     }
                     if (owner!=-1)
